Update: Define public interface for package `logger`
This commit adds public functions `logger.Log` and `logger.Stop`

The function `logger.Log` switches the logger to log events at
debug level, or above (as opposed to the `warn` level by default).
The function can optionally enable writing logs to a log file
if needed

The function `logger.Stop`, as the name implies, stops the logger.
It aims to gracefully close the logger, especially when writing to
a log file - ensuring that the file stream closes only after writing
all contents to the log file

Signed-off-by: notsatan <git@notsatan.me>