Feature: Enable logging when `debug` mode is detected
The package `logger` will now switch to logging events at debug
mode or above anytime the application detects it is running in
debug mode

For context: logging is always enabled, but the logger ignores
events below `warn` level. That is, the logger ignores messages
that are not a warning or an error. On the other hand, running the
function `logger.Log` - switches the logger to log messages at
`debug` level or above

Signed-off-by: notsatan <git@notsatan.me>