Refactor: Merge package `src` into package `src/cmd`
This commit merges the existing code from `src/src.go` into
`src/cmd.go` - almost a rewrite of notsatan/crcgen@a93a95007e8a3f

In certain scenario's, the root command would need to be able to
force-close the flow directly. As such, it won't be possible to
return errors back to the calling functions to form a stack-trace

To ensure a safe exit, a call to `src.closeRes` would be needed,
which means that the scope of the function required to be public.

Instead of attempting to modify the function to be public, at which
point, code in the module `src` would have no real use - merging
the existing package `src` into `cmd` seemed to be a better option.

No real changes were made to the code except for moving stuff around,
and so - which is why this change was a simple commit and not a
refactor branch :p

Signed-off-by: notsatan <git@notsatan.me>