Update: Add generic functions to deal with filepath
Define package `src/lib` to contain general modules that can
be freely used across the codebase

Add new generic functions to deal with filesystems;
    func IsInvalidPathErr(error) bool
    func PathExists(string) bool

Define a local error
    var errInvalidPath error

The error is used to indicate an invalid file path. This will be
used in a future commit.

The function `IsInvalidPathErr` determines if an error was caused
because of invalid input path, while `PathExists` acts as a helper
function to detect if a filepath is valid or not

Signed-off-by: notsatan <git@notsatan.me>