Update: Define function `lib.WalkPath` to walk directories
This commits adds a generic function `lib.WalkPath` that traverses
the root directory, iterating through all the files under it.

The function is a wrapper over `filepath.Walk` - taking a function
as an input - this function will be run over each file present
anywhere within the root directory.
    func WalkPath(path string, walkFunc filepath.WalkFunc) error

The function `WalkPath` includes a custom check to validate the
root directory, if the path is incorrect, an `errInvalidPath`
error will be returned by the function

`WalkPath` will act as the single flow-of-control regarding files
read by `crcgen` - i.e. if files or directories are to ignored in
the future, it will be as simple as modifying this function

Signed-off-by: notsatan <git@notsatan.me>