Update: Return error for output files without an extension
The previous commit assumed that a file without a valid extension
should just default to JSON. After putting some thought to the
same, it was decided to revert that - with this commit, file
paths without a valid extension will return an inavlid extension
error

Currently, the plan is to point viper directly to the file using
the path instead of allowing viper to locate the file by itself.
Ensuring file extensions are a part of the filename is a part of
this move

This commit modifies the codebase, as well as the related test(s).

Given that this change is being made to a development branch - this
won't be considered a breaking change, and instead would be a
simple patch

Signed-off-by: notsatan <git@notsatan.me>