Update: Check to ensure output file exists, create one if needed
Up until now, the code was simply performing checks based on the
full path, filename and extension - no checks were being made to
ensure that the file being used as output file actually existed

This commit adds a check to verify if the output file exists, if
not, an empty file is created. Added two new sentinel errors
`errPathIsDir` and `errNotWritable`

The former indicates that the output path points to an existing
directory, and thus cannot be used as the output file. For example
the path `/root/user/output.json` can belong to a directory, or a
file named `output.json` in Unix

The latter error indicates that the path to output file is not
write-able, currently, if the file creation failed for any reason,
the app would assume that the cause was missing write perms, and
throw this error (this can be polished further in the future)

Both these sentinel errors are local to the package, and can be
tested against using the corresponding public functions

While creating a new output file, the program attempts to handle
some corner cases as well;
 - Check to verify if the path is occupied
 - Ensure that the path points to a file (and not a directory)
 - If any error is caused during this check, fail the process
 - If the path is empty, proceed to create an empty file

Signed-off-by: notsatan <git@notsatan.me>