Update: Add method `DirInfo.CalcModTime`
In some cases, OS'es do not store last modtime for directories, as
such, it can't be assumed that `DirInfo` objects would always have
a value against `LastMod` time

The method `CalcModTime` aims to resolve this, for objects that
have a value against `DirInfo.LastMod`, the method directly returns
the existing value

On the other hand, for objects without a set a value, the method
attempts to calculate a value based on the last modification time
for all files under that directory in the file tree - including
files in sub-directories!

The method makes an internal recursive call for each directory,
and uses the value returned. Since the last mod time is a measure
of epoch seconds when the file/directory was last modified, the
greatest value among all child items will be picked as the last
mod time for the current directory

Signed-off-by: notsatan <git@notsatan.me>