Fix: Patch crash caused by nil map declaration
The declaration of `outHandlers` in `writers/handler.go` simply
declared a map, without instantiating it - as a result, the map would
remain nil

Any attempts to add values to the nil map would result in a crash,
which is what would happen at any attempt to run the script. This
commit simply fixes the cause of by instantiating `outHandlers`
when declared

P.S. As to why this "bug" went undetected in tests? That's because
the helper test function `writer.reset()` would reset the value of
the map (and other variables) - and while doing so, it would
instantiate the map as well

Signed-off-by: notsatan <git@notsatan.me>