Update: Define method `writer.Write` to write `DirInfo` objects
This commit introduces the method `writer.Write` that allows writing
a `DirInfo` object to the output file

    func Write(info *DirInfo) error

When called, the function internally fethes an appropriate handler
based on output file extension, marshals the `DirInfo` object that
is to be written, and then writes the same into the output file

If at any step of the process, an error is encountered, the function
will simply return an error indicating failure. In case of success,
all existing content of the file will be overwritten with the data

Created a new structure `mockHandlerFail` for tests, this is similar
to `mockHandler`, but fails every method where possible. Ensuring
greater test coverage more scenarios can be emulated

Signed-off-by: notsatan <git@notsatan.me>